Windows: dummy mount
