+++ mount_windows.go
@@ -1,0 +1,9 @@
+package containerd
+
+import (
+	"errors"
+)
+
+func (m *Mount) Mount(target string) error {
+	return errors.New("mount not supported")
+}
